Relocate ClaimAnimation
